- Moved legacy method tests to PegUtilsLegacyTest. - Updated other test places where legacy methods were being used.
